requirements.txt
requests
beautifulsoup4
lxml

langchain
langchain-core
langchain-groq
langchain-openai
langchain-community
pydantic

python-dotenv

Flask
Flask-Cors